Move the history tokenizer classes to a lib so that it could be extracted to a gem
