+++ lib/historical.rb
@@ -1,0 +1,10 @@
+require_relative "historical/actor_history_token"
+require_relative "historical/change_history_token"
+require_relative "historical/created_history_token"
+require_relative "historical/date_history_token"
+require_relative "historical/event_history_token"
+require_relative "historical/history_token_parser"
+require_relative "historical/history_token_registry"
+require_relative "historical/history_tokenizer"
+require_relative "historical/registered_token"
+require_relative "historical/time_history_token"
