+++ lib/historical/actor_history_token.rb
