+++ lib/historical/created_history_token.rb
