+++ lib/historical/date_history_token.rb
