+++ lib/historical/event_history_token.rb
