+++ lib/historical/history_token_parser.rb
