+++ lib/historical/history_token_registry.rb
