+++ lib/historical/history_tokenizer.rb
