+++ lib/historical/registered_token.rb
