+++ lib/historical/time_history_token.rb
